Refactors python codebase
